+++ src/main/java/org/openlmis/hapifhir/Application.java
@@ -16,7 +16,6 @@
 package org.openlmis.hapifhir;
 
 import java.util.Locale;
-
 import org.openlmis.hapifhir.i18n.ExposedMessageSourceImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -24,7 +23,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.flyway.FlywayMigrationStrategy;
-import org.springframework.boot.web.servlet.ServletComponentScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ImportResource;
 import org.springframework.context.annotation.Profile;
@@ -33,7 +31,6 @@ import org.springframework.web.servlet.i18n.CookieLocaleResolver;
 
 
 @SpringBootApplication
-@ServletComponentScan
 @ImportResource("applicationContext.xml")
 public class Application {