+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 1.0.0 / WIP
 ==================
 
-[OLMIS-5382](https://openlmis.atlassian.net/browse/OLMIS-5382) Create a HAPI FHIR microservice.
+Released openlmis-hapifhir 1.0.0 as part of openlmis-ref-distro 3.5. This was the first stable release of openlmis-hapifhir.
+
+Features:
+* [OLMIS-5382](https://openlmis.atlassian.net/browse/OLMIS-5382): Create a HAPI FHIR microservice.
+* [OLMIS-5385](https://openlmis.atlassian.net/browse/OLMIS-5385): Supports only service-based or API key tokens